third_party/WebKit/LayoutTests/fast/css/focus-ring-recursive-continuations-expected.txt

When adding focus ring rects, we should avoid adding duplicated rect for continuations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
FAIL internals.focusRingRects(document.getElementById('focus')) should be 5 (of type number). Was [object ClientRectList] (of type object).
SPAN0
DIV1 SPAN1
DIV2
